fix(functions/api): combine smaller fetch funcs into one file and make suggestions behave as others
