+++ src/lib/functions/api/fetchapi.js
@@ -1,100 +1,47 @@
-import { error } from '@sveltejs/kit';
-import { createApiUrl } from './createurl';
+import { concatSearchParams } from './concatparams';
+import { fetchAPI } from './fetchapigeneric';
 
 /**
- * @template T
- * @param {typeof fetch} fetcher
- * @param {string} path
- * @param {URLSearchParams} [params]
- * @returns {Promise<T>}
+ * Fetches /currencies from the API.
+ * @param {typeof fetch} [fetcher]
+ * @returns {Promise<CurrenciesResponseType>}
  */
-export async function fetchAPI(fetcher, path, params) {
-	/** @type {URL} */
-	let apiUrl;
-	try {
-		apiUrl = createApiUrl(path, params);
-	} catch (err) {
-		// Internal Server Error.
-		throw error(500, `Failed to create API URL: ${err}`);
-	}
-
-	/** @type {Response} */
-	let response;
-	try {
-		// Fetch the results from the API.
-		response = await fetcher(apiUrl, {
-			method: 'GET', // POST doesn't cache on CDN.
-			headers: {
-				Accept: 'application/json',
-				'Accept-Encoding': 'gzip, deflate, br'
-			}
-		});
-	} catch (err) {
-		// Bad Gateway.
-		throw error(502, `Failed to fetch: ${err}`);
-	}
-
-	// Check if the response is OK, otherwise handle the error.
-	if (!response.ok) {
-		// Try to read the response body as text.
-		let responseBody;
-		try {
-			responseBody = await response.text();
-		} catch (err) {
-			responseBody = `Failed to read response body: ${err}`;
-		}
-
-		// Same as the API.
-		throw error(
-			response.status,
-			`Response from API not OK (${response.statusText}): ${responseBody}`
-		);
-	}
-
-	/** @type {T | ErrorResponseType} */
-	let jsonResponse;
-	try {
-		// Parse the response body as JSON.
-		jsonResponse = await response.json();
-	} catch (err) {
-		/** @type {string} */
-		let textResponse;
-		try {
-			// Try to read the response body as text.
-			textResponse = await response.text();
-		} catch (textErr) {
-			// Bad Gateway.
-			throw error(
-				502,
-				`Failed to read response body: ${textErr} (${response.status} ${response.statusText}): ${err}`
-			);
-		}
-
-		// Internal Server Error.
-		throw error(
-			500,
-			`Failed to parse JSON body: ${err} (${response.status} ${response.statusText}): ${textResponse}`
-		);
-	}
-
-	// Check if the response is an error from the API (ErrorResponseType).
-	if (isErrorResponse(jsonResponse)) {
-		// Same as the API.
-		throw error(
-			response.status,
-			`Response from API is an error: ${jsonResponse.message}: ${jsonResponse.value}`
-		);
-	}
-
-	return jsonResponse;
+export async function fetchCurrencies(fetcher = fetch) {
+	return await fetchAPI(fetcher, 'currencies');
 }
 
 /**
- * Type guard to check if the response is an ErrorResponseType.
- *
- * @param {any} response
- * @returns {response is ErrorResponseType}
+ * Fetches /exchange with params from the API.
+ * @param {URLSearchParams} params
+ * @param {typeof fetch} [fetcher]
+ * @returns {Promise<ExchangeResponseType>}
  */
-function isErrorResponse(response) {
-	return typeof response === 'object' && 'message' in response && 'value' in response;
+export async function fetchExchange(params, fetcher = fetch) {
+	return await fetchAPI(fetcher, 'exchange', params);
+}
+
+/**
+ * Fetches /search with params from the API.
+ * @param {URLSearchParams} params
+ * @param {typeof fetch} [fetcher]
+ * @returns {Promise<ResultsResponseType>}
+ */
+export async function fetchResults(params, fetcher = fetch) {
+	return await fetchAPI(fetcher, 'search', params);
+}
+
+/**
+ * Fetches /suggestions with query from the API.
+ * @param {string} query
+ * @param {typeof fetch} [fetcher]
+ * @returns {Promise<SuggestionsResponseType>}
+ */
+export async function fetchSuggestions(query, fetcher = fetch) {
+	/** @type {URLSearchParams} */
+	const params = concatSearchParams([
+		['q', query],
+		['output', 'json']
+	]);
+
+	return await fetchAPI(fetcher, 'suggestions', params);
 }
